[int] Use ‘off_t’ for ‘behavior.mem_limit’.
Incidentally, this eliminates a "comparison between signed
and unsigned integer expressions" warning from GCC 4.4.

* src/base.h (struct behavior): Change type of ‘mem_limit’ to ‘off_t’.
* src/b-fro.c (fro_open): Delete cast to ‘off_t’.

+++ src/base.h
@@ -515,7 +515,7 @@ struct behavior
   /* The current working directory.
      -- getfullRCSname  */
 
-  size_t mem_limit;
+  off_t mem_limit;
   /* If a fro is smaller than ‘mem_limit’ kilobytes, try to mmap(2) it
      (if mmap(2)), or operate on a copy of it in core (if no mmap(2)).
      Otherwise, use standard i/o routines as the fallback.
